enemy wave stuff

+++ RotoShootUnityProject/Assets/Scripts/LevelSetupData.cs
@@ -6,7 +6,7 @@ public class LevelSetupData : ScriptableObject
 {
   public int levelNumber = 0;
   public int levelControlType = 2;
-  public float numMaxActiveWaves;
+  public int numMaxActiveWaves;
   public float minTimeBetweenWaves, maxTimeBetweenWaves;
 
   public EnemySpawnPointData[] levelEnemySpawnPointData;
